requirements.txt
SQLAlchemy~=1.3.17
pandas~=1.0.3
PyYAML~=5.3.1
Deprecated~=1.2.10
twine~=3.1.1
wheel~=0.34.2